Expose complete history to QML
The use case for this functionality is to be able to e.g. display the terminal application's output after it was terminated.
This requires exposing a `clearScreen` method to QML as the history only contains the lines not currently displayed in the widget. If we want to access the whole output we have to be able to flush the screen beforehand.

The single argument overload of `Emulation::writeToStream` simplifies reading the complete history at once without having to further expose it's length.