Take transient scrollbars into account
Fixes https://github.com/lxqt/qterminal/issues/415

Because they are transient by definition (since Qt 5.5), the code should not spare an extra space for them (Konsole does not have this feature).

Apart from that, the code does not need to fill the area behind the scrollbar explicitly; it is enough to auto-fill it.

Naturally, no change will be seen with styles other than Kvantum.