enable bidi rendering by default (#481)
Arabic, Persian, Urdu, Hebrew, and many other languages are written in
an RTL script. And, naturally, people prefer to have filenames etc in
their native language, regardless of the UI language.

Konsole enables BiDi rendering by default, and for a good reason.

As this is a library, I think it needs this to be enabled by default
more than a stand-alone application like Konsole, because it's very easy
for an application to overlook this feature, and not even provide an
option to enable it, rendering it much less useful than it would have
been.