Assign bulkTimers a parent
The parent/child relationship is not used to manage memory.
It's only used to keep the objects together when moving them between
threads (so that the entire subtree is moved by moveToThread).
That's not the case here, but it's good practice.

Reference: https://www.kdab.com/qobject_ownership_propagate_const/