Fixed URL highlighting on wrapping
Previously, if a line consisted of two URLs separated by a space and was wrapped at the end of that space, both URLs were underlined if the cursor was on each of them.

An example of wrapping by resizing QTerminal (for testing):

```
abc://aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
abc://aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
```

Note that there is a space at the end of the first line.

Fixes https://github.com/lxqt/qterminal/issues/1184