Support combining characters (#571)
I consulted the code of the latest Konsole to add support for combining characters to qtermwidget. It was a step-by-step process of testing and bug fixing. Therefore, all parts of this huge change are required.

Closes https://github.com/lxqt/qterminal/issues/319

NOTE: The patch passed my tests, but *please do not merge the PR unless it's either tested thoroughly or reviewed.*