Make too wide characters fit into available space
The drawing is done like that of XTerm (actually, Qt does it), i.e., the character is "squeezed". A too wide character is defined by the code as a character whose width is twice the monospace width or more. An example is "three-em dash", which is `⸻`.

Fixes https://github.com/lxqt/qterminal/issues/685