Support emoji flags
Because emoji flags are pairs of "regional indicators", they're treated as a special case of extended characters.

NOTE: A few nonstandard flags, that are created by the so-called "emoji tag sequences", aren't supported by this commit. Since this small commit was the result of several hours of work, I'm not motivated to spend more time on adding features like this. However, clean PRs would be appreciated and reviewed.

Closes https://github.com/lxqt/qterminal/issues/1268